+++ src/types/index.ts
@@ -11,8 +11,9 @@ export interface ChatProviderType {
 export interface ChatContextType extends BlockListHook, UsersHookExports, NetworkStatusHook {
   isConnected: boolean;
   chatStatus: ChatStatus;
-  connect: (credentials: Chat.ConnectionParams) => Promise<void>;
-  disconnect: () => void;
+  connect: (credentials: Chat.ConnectionParams) => Promise<boolean>;
+  disconnect: () => Promise<boolean>;
+  terminate: () => void;
   currentUserId?: number;
   createChat: (userId: number, extensions?: { [key: string]: any }) => Promise<Dialogs.Dialog>;
   createGroupChat: (
@@ -22,6 +23,8 @@ export interface ChatContextType extends BlockListHook, UsersHookExports, Networ
     extensions?: { [key: string]: any },
   ) => Promise<Dialogs.Dialog>;
   getDialogs: (filters?: Dialogs.ListParams) => Promise<Dialogs.Dialog[]>;
+  getNextDialogs: () => Promise<Dialogs.Dialog[]>;
+  totalDialogReached: boolean;
   dialogs: Dialogs.Dialog[];
   selectedDialog?: Dialogs.Dialog;
   selectDialog: (dialog?: Dialogs.Dialog) => Promise<void>;
@@ -29,6 +32,7 @@ export interface ChatContextType extends BlockListHook, UsersHookExports, Networ
   unreadMessagesCount: { total: number; [dialogId: string]: number };
   getMessages: (dialogId: string) => Promise<Messages.Message[]>;
   getNextMessages: (dialogId: string) => Promise<Messages.Message[]>;
+  totalMessagesReached: { [dialogId: string]: boolean };
   messages: { [key: string]: Messages.Message[] };
   markDialogAsRead: (dialog: Dialogs.Dialog) => Promise<void>;
   addUsersToGroupChat: (usersIds: number[]) => Promise<void>;
@@ -48,7 +52,7 @@ export interface ChatContextType extends BlockListHook, UsersHookExports, Networ
   processOnMessageSent: (fn: Chat.OnMessageSentListener | null) => void;
 }
 
-export enum GroupChatEventType {
+export enum DialogEventSignal {
   ADDED_TO_DIALOG = "dialog/ADDED_TO_DIALOG",
   REMOVED_FROM_DIALOG = "dialog/REMOVED_FROM_DIALOG",
   ADD_PARTICIPANTS = "dialog/ADD_PARTICIPANTS",